+++ lib/jitsi_meeting_listener.dart
@@ -1,10 +1,8 @@
-import 'package:flutter/foundation.dart';
-
 /// Class holding the callback functions for conference events
 class JitsiMeetingListener {
-  final VoidCallback onConferenceWillJoin;
-  final VoidCallback onConferenceJoined;
-  final VoidCallback onConferenceTerminated;
+  final Function({Map<dynamic, dynamic> message}) onConferenceWillJoin;
+  final Function({Map<dynamic, dynamic> message}) onConferenceJoined;
+  final Function({Map<dynamic, dynamic> message}) onConferenceTerminated;
   final Function(dynamic error) onError;
 
   JitsiMeetingListener(